requirements.txt
# Core ML/DL (Updated for Python 3.12 compatibility)
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.40.0
torch-geometric>=2.5.0
sentence-transformers>=2.7.0

# Explainability (NEW)
captum>=0.7.0  # Integrated Gradients
shap>=0.45.0
lime>=0.2.0.1

# Code Analysis
tree-sitter>=0.23.0
tree-sitter-c>=0.24.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0

# Graph Database (NEW)
neo4j>=5.19.0
py2neo>=2021.2.4

# Local Agent (NEW)
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9

# Symbolic Execution (NEW) - Note: angr may have Python 3.12 compatibility issues
# angr>=9.2.100; python_version < '3.12'  # Commented out for Python 3.12
z3-solver>=4.13.0

# Data & Processing
pandas>=2.2.0
numpy>=1.26.4
scipy>=1.13.0
faiss-cpu>=1.8.0

# API & Serialization
pydantic>=2.7.0
python-jsonrpc-server>=0.4.0

# Development
pytest>=8.1.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
black>=24.3.0
pylint>=3.1.0
mypy>=1.9.0
isort>=5.13.2

# AWS
boto3>=1.34.0
sagemaker>=2.214.0

# Utilities
python-dotenv>=1.0.1
rich>=13.7.1
click>=8.1.7
watchdog>=4.0.0  # File monitoring

# Additional dependencies for Git operations
GitPython>=3.1.43
requests>=2.31.0
